requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
arxiv==1.4.7
pandas==1.5.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
scipy==1.10.1
tqdm==4.65.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
python-dotenv==1.0.0 